+++ hikka/modules/loader.py
@@ -1011,8 +1011,11 @@ class LoaderMod(loader.Module):
 
                 return await self.load_module(**kwargs)  # Try again
             except loader.LoadError as e:
-                with contextlib.suppress(ValueError):
-                    self.allmodules.modules.remove(instance)  # skipcq: PYL-E0601
+                with contextlib.suppress(Exception):
+                    self.allmodules.unload_module(instance.__class__.__name__)
+
+                with contextlib.suppress(Exception):
+                    self.allmodules.modules.remove(instance)
 
                 if message:
                     await utils.answer(
@@ -1078,7 +1081,10 @@ class LoaderMod(loader.Module):
                 )
                 task.cancel()
             except loader.LoadError as e:
-                with contextlib.suppress(ValueError):
+                with contextlib.suppress(Exception):
+                    self.allmodules.unload_module(instance.__class__.__name__)
+
+                with contextlib.suppress(Exception):
                     self.allmodules.modules.remove(instance)
 
                 if message:
@@ -1090,7 +1096,10 @@ class LoaderMod(loader.Module):
                 return
             except loader.SelfUnload as e:
                 logging.debug(f"Unloading {instance}, because it raised `SelfUnload`")
-                with contextlib.suppress(ValueError):
+                with contextlib.suppress(Exception):
+                    self.allmodules.unload_module(instance.__class__.__name__)
+
+                with contextlib.suppress(Exception):
                     self.allmodules.modules.remove(instance)
 
                 if message: